requirements.txt
django-environ==0.10.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0
